requirements.txt
click==7.1.2
flask==1.1.2
flask-sqlalchemy==2.4.3
itsdangerous==1.1.0
jinja2==2.11.2
joblib==0.15.1
markupsafe==1.1.1
numpy==1.19.0
pandas==1.0.5
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
scikit-surprise==1.1.0
scipy==1.5.0
six==1.15.0
sqlalchemy==1.3.18
werkzeug==1.0.1
